requirements.txt
torch
transformers
datasets
evaluate
jiwer
openai-whisper
sounddevice
numpy
pandas
scipy